requirements.txt
<<<<<<< HEAD
flask==3.1.0
groq==0.25.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.4.1
PyPDF2==3.0.1
python-dotenv==1.0.1
werkzeug==3.1.3
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
langchain-mistralai>=0.2.0
watchdog>=6.0.0
python-pptx
openpyxl

=======
streamlit
langchain
langchain-core
langchain-text-splitters
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
PyPDF2
python-docx
python-dotenv
sentence-transformers
>>>>>>> 5caf138cd4bf724bc60bc7be180102528e7f7762
